Understand no version as latest and install over
When no version is specified, this will be understood as latest. This
also means that `primer` will always look for the latest version and
install the latest version of a tool, even though it is already
installed. In other words: the default is to upgrade when the version is
empty.

When a version is specified, `primer` will try to satisfy that version,
even though the tool is already installed. Thus this might upgrade, but
also downgrade.